+++ src/client.js
@@ -639,7 +639,7 @@ Client.prototype.telemetryMonitor = null;
  * @static
  * @type {String}
  */
-Client.version = '3.3.1';
+Client.version = '3.3.2';
 
 /**
  * Any Conversation or Message that is part of a Query's results are kept in memory for as long as it